refactor: moved default_invoke_tx_args to be associated with InvokeTxArgs
